Remove campaigns folder
