+++ drip/templates/admin/drip/Campaign/change_form.html
@@ -5,7 +5,7 @@
   >
 </li>
 <li>
-  <a href="{% url 'admin:campaigns_campaign_history' original.pk %}" class="historylink">{% translate "History" %}</a>
+  <a href="{% url 'admin:drip_campaign_history' original.pk %}" class="historylink">{% translate "History" %}</a>
 </li>
 {% if has_absolute_url %}
 <li>
