+++ drip/templates/campaign/timeline.html
